- Refactoring and improving the BR module
