+++ src/community/backup-restore/core/src/main/java/org/geoserver/backuprestore/AbstractExecutionAdapter.java
@@ -4,20 +4,11 @@
  */
 package org.geoserver.backuprestore;
 
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.Collections;
-import java.util.Date;
-import java.util.List;
+import java.util.*;
 import java.util.concurrent.ConcurrentHashMap;
 import org.geoserver.platform.resource.Resource;
 import org.opengis.filter.Filter;
-import org.springframework.batch.core.BatchStatus;
-import org.springframework.batch.core.ExitStatus;
-import org.springframework.batch.core.JobExecution;
-import org.springframework.batch.core.JobInstance;
-import org.springframework.batch.core.JobParameters;
-import org.springframework.batch.core.StepExecution;
+import org.springframework.batch.core.*;
 
 /**
  * Base Class for {@link JobExecution} wrappers. Those will be used to share objects, I/O parameters
