+++ src/community/backup-restore/core/src/test/java/org/geoserver/backuprestore/BackupWithoutSettingsTest.java
@@ -1,8 +1,6 @@
 package org.geoserver.backuprestore;
 
-import static org.junit.Assert.assertNotNull;
-import static org.junit.Assert.assertNull;
-import static org.junit.Assert.assertTrue;
+import static org.junit.Assert.*;
 
 import java.io.File;
 import java.util.HashMap;
