+++ src/community/backup-restore/core/src/test/java/org/geoserver/backuprestore/GwcBackupTest.java
@@ -4,9 +4,7 @@
  */
 package org.geoserver.backuprestore;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNotNull;
-import static org.junit.Assert.assertTrue;
+import static org.junit.Assert.*;
 
 import java.io.File;
 import java.util.HashMap;
