+++ src/community/backup-restore/core/src/test/java/org/geoserver/backuprestore/GwcRestoreTest.java
@@ -4,9 +4,7 @@
  */
 package org.geoserver.backuprestore;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNotNull;
-import static org.junit.Assert.assertTrue;
+import static org.junit.Assert.*;
 
 import java.util.HashMap;
 import java.util.Map;
