+++ src/community/backup-restore/core/src/test/resources/applicationContext.xml
@@ -5,24 +5,17 @@
  This code is licensed under the GPL 2.0 license, available at the root
  application directory. 
 -->
-<beans xmlns="http://www.springframework.org/schema/beans"
-	xmlns:batch="http://www.springframework.org/schema/batch"
-	xmlns:util="http://www.springframework.org/schema/util"
-	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:schemaLocation="
-		http://www.springframework.org/schema/beans
-        http://www.springframework.org/schema/beans/spring-beans-4.0.xsd
-        http://www.springframework.org/schema/batch
-        http://www.springframework.org/schema/batch/spring-batch-3.0.xsd
-        http://www.springframework.org/schema/util
-        http://www.springframework.org/schema/util/spring-util-4.0.xsd">
+<beans xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+       xmlns="http://www.springframework.org/schema/beans"
+       xsi:schemaLocation="http://www.springframework.org/schema/beans
+        http://www.springframework.org/schema/beans/spring-beans-4.0.xsd">
 
-	<bean id="extraFileHandler" class="org.geoserver.backuprestore.ExtraFileHandler">
-		<constructor-arg ref="dataDirectory"/>
-	</bean>
+    <bean id="extraFileHandler" class="org.geoserver.backuprestore.ExtraFileHandler">
+        <constructor-arg ref="dataDirectory"/>
+    </bean>
 
-	<bean id="continuableHandler" class="org.geoserver.backuprestore.ContinuableHandler"/>
+    <bean id="continuableHandler" class="org.geoserver.backuprestore.ContinuableHandler"/>
 
-	<bean id="genericListener" class="org.geoserver.backuprestore.GenericListener"/>
+    <bean id="genericListener" class="org.geoserver.backuprestore.GenericListener"/>
 
 </beans>
