+++ src/community/backup-restore/extension/src/main/resources/applicationContext.xml
@@ -5,19 +5,19 @@
 	directory. -->
 
 <beans>
-	<bean class="org.geoserver.platform.ModuleStatusImpl">
-		<constructor-arg index="0" value="gs-backup-restore-extension," />
-		<constructor-arg index="1" value="Backup Rstore Extensions" />
-	</bean>
+    <bean class="org.geoserver.platform.ModuleStatusImpl">
+        <constructor-arg index="0" value="gs-backup-restore-extension,"/>
+        <constructor-arg index="1" value="Backup Rstore Extensions"/>
+    </bean>
 
-	<!-- CatalogAdditionalResourcesWriter extensions: those can be triggered by the CatalogWriter -->
-	
-	<bean id="imageMosaicAdditionalResourceReader"
-		class="org.geoserver.backuprestore.imagemosaic.reader.ImageMosaicAdditionalResourceReader" />
+    <!-- CatalogAdditionalResourcesWriter extensions: those can be triggered by the CatalogWriter -->
 
-	<!-- CatalogAdditionalResourcesReader extensions: those can be triggered by the CatalogReader -->
-	
-	<bean id="imageMosaicAdditionalResourceWriter"
-		class="org.geoserver.backuprestore.imagemosaic.writer.ImageMosaicAdditionalResourceWriter" />
+    <bean id="imageMosaicAdditionalResourceReader"
+          class="org.geoserver.backuprestore.imagemosaic.reader.ImageMosaicAdditionalResourceReader"/>
+
+    <!-- CatalogAdditionalResourcesReader extensions: those can be triggered by the CatalogReader -->
+
+    <bean id="imageMosaicAdditionalResourceWriter"
+          class="org.geoserver.backuprestore.imagemosaic.writer.ImageMosaicAdditionalResourceWriter"/>
 
 </beans>
