+++ src/community/backup-restore/web/src/main/java/org/geoserver/backuprestore/web/StoreModel.java
@@ -14,10 +14,9 @@ import org.geoserver.web.GeoServerApplication;
 public class StoreModel<T extends StoreInfo> extends LoadableDetachableModel<T> {
 
     private static final long serialVersionUID = -2008934085507417813L;
-
-    private ResourceFilePanel resourceFilePanel;
     IModel workspace;
     String name;
+    private ResourceFilePanel resourceFilePanel;
 
     private StoreModel(T store) {
         super(store);
