+++ src/community/backup-restore/web/src/main/resources/GeoServerApplication.properties
@@ -5,7 +5,6 @@ done=Done
 apply=Apply
 refresh=Refresh
 mandatory=Mandatory Fields
-
 BackupRestoreDataPage.menu=Backup & Restore
 BackupRestoreDataPage.title=Backup & Restore
 BackupRestoreDataPage.description=
@@ -32,19 +31,16 @@ BackupRestoreDataPage.store=Filter by Store (optional)
 BackupRestoreDataPage.store.nullValid=(all)
 BackupRestoreDataPage.layer=Filter by Layer (optional)
 BackupRestoreDataPage.layer.nullValid=(all)
-
 BackupRestoreExecutionsTable.th.id=Execution
 BackupRestoreExecutionsTable.th.state=State
 BackupRestoreExecutionsTable.th.started=Created
 BackupRestoreExecutionsTable.th.options=Options
 BackupRestoreExecutionsTable.th.progress=Progress
 BackupRestoreExecutionsTable.th.archiveFile=Archive File
-
 ResourceFilePanel.browse=Browse
 ResourceFilePanel.file=Provide target archive full path
 ResourceFilePanel.submit=Submit
 ResourceFilePanel.chooseFile=Provide target archive full path
-
 BackupRestorePage.title=
 BackupRestorePage.description=
 BackupRestorePage.execution=Execution
