+++ src/community/backup-restore/web/src/main/resources/GeoServerApplication_ko.properties
@@ -5,7 +5,6 @@ done=\uc644\ub8cc
 apply=\uc801\uc6a9
 refresh=\uc0c8\ub85c\uace0\uce68
 mandatory=\ud544\uc218 \ud544\ub4dc
-
 BackupRestoreDataPage.menu=\ubc31\uc5c5 \ubc0f \ubcf5\uc6d0
 BackupRestoreDataPage.title=\ubc31\uc5c5 \ubc0f \ubcf5\uc6d0
 BackupRestoreDataPage.description=
@@ -32,19 +31,16 @@ BackupRestoreDataPage.store=\uc800\uc7a5\uc18c\ub85c \ud544\ud130\ub9c1 (\uc635\
 BackupRestoreDataPage.store.nullValid=(\ubaa8\ub450)
 BackupRestoreDataPage.layer=\ub808\uc774\uc5b4\ub85c \ud544\ud130\ub9c1 (\uc635\uc158)
 BackupRestoreDataPage.layer.nullValid=(\ubaa8\ub450)
-
 BackupRestoreExecutionsTable.th.id=\uc2e4\ud589
 BackupRestoreExecutionsTable.th.state=\uc0c1\ud0dc
 BackupRestoreExecutionsTable.th.started=\uc0dd\uc131\ub428
 BackupRestoreExecutionsTable.th.options=\uc635\uc158
 BackupRestoreExecutionsTable.th.progress=\uc9c4\ud589
 BackupRestoreExecutionsTable.th.archiveFile=\uc544\uce74\uc774\ube0c \ud30c\uc77c
-
 ResourceFilePanel.browse=\ud0d0\uc0c9
 ResourceFilePanel.file=\ub300\uc0c1 \uc544\uce74\uc774\ube0c \uc804\uccb4 \uacbd\ub85c \uc81c\uacf5
 ResourceFilePanel.submit=\uc81c\ucd9c
 ResourceFilePanel.chooseFile=\ub300\uc0c1 \uc544\uce74\uc774\ube0c \uc804\uccb4 \uacbd\ub85c \uc81c\uacf5
-
 BackupRestorePage.title=
 BackupRestorePage.description=
 BackupRestorePage.execution=\uc2e4\ud589
